Reorganize the code
Previously, the code assumes that it would try to index all media files
in the current working directory, and index them separately for
different media types.

This works, until a Dropbox-like service is planned to be introduced.
It would be hard to infer which file is intended to be indexed for
download.  As a result, a JSON-based config file is introduced to
fine-tune the indexing for different indexers.